+++ config/platform.php
@@ -156,7 +156,9 @@ return [
     | Example: [ 'fa' => storage_path('app/fontawesome') ]
     */
 
-    'icons' => [],
+    'icons' => [
+        'orc' => \Orchid\IconPack\Path::getFolder(),
+    ],
 
     /*
     |--------------------------------------------------------------------------