+++ resources/js/controllers/upload_controller.js
@@ -1,7 +1,6 @@
 import ApplicationController from "./application_controller"
 import {Dropzone} from 'dropzone';
 import Sortable from 'sortablejs';
-import {debounce, has as objHas} from "lodash";
 
 export default class extends ApplicationController {
 
@@ -16,6 +15,7 @@ export default class extends ApplicationController {
         'alt',
         'description',
         'url',
+        'loadmore',
     ];
 
     /**
@@ -25,11 +25,12 @@ export default class extends ApplicationController {
     constructor(props) {
         super(props);
         this.attachments = {};
-        this.mediaList = {};
+        this.mediaList = [];
+        this.allMediaList = {};
     }
 
     initialize() {
-        this.loadMedia = debounce(this.loadMedia, 500);
+        this.page = 1
     }
 
     /**
@@ -263,7 +264,7 @@ export default class extends ApplicationController {
                 });
 
                 this.on('removedfile', file => {
-                    if (objHas(file, 'data.id')) {
+                    if (file.hasOwnProperty('data.id')) {
                         $(dropname).find(`.files-${file.data.id}`).remove();
                         !isMediaLibrary && axios
                             .delete(urlDelete + file.data.id, {
@@ -343,6 +344,25 @@ export default class extends ApplicationController {
     /**
      *
      */
+    loadMore(event) {
+        event.preventDefault();
+        this.page++;
+        this.loadMedia();
+    }
+
+    /**
+     *
+     */
+    resetPage() {
+        this.allMediaList = {}; // Reset all media list
+        this.page = 1; // Reset page
+
+        $(this.dropname).find(`.media-results`).html('');
+    }
+
+    /**
+     *
+     */
     loadMedia() {
         const self = this;
         const CancelToken = axios.CancelToken;
@@ -350,10 +370,11 @@ export default class extends ApplicationController {
         if (typeof this.cancelRequest === 'function') {
             this.cancelRequest();
         }
+
         $(this.dropname).find(`.media.modal`).modal('show');
 
         axios
-            .post(this.prefix('/systems/media'), {
+            .post(this.prefix(`/systems/media?page=${this.page}`), {
                 filter: {
                     disk: this.data.get('storage'),
                     original_name: this.searchTarget.value,
@@ -364,7 +385,9 @@ export default class extends ApplicationController {
                 }),
             })
             .then((response) => {
-                this.mediaList = response.data;
+                this.mediaList = response.data.data;
+                // show/hide load more
+                this.loadmoreTarget.classList.toggle('d-none', response.data.last_page === this.page);
                 this.renderMedia();
             });
     }
@@ -373,21 +396,25 @@ export default class extends ApplicationController {
      *
      */
     renderMedia() {
-        let html = '';
-
-        /** todo: */
         this.mediaList.forEach((element, key) => {
-            html += '<div class="col-4 col-sm-3 my-3 position-relative media-item">\n' +
-                '    <div data-action="click->upload#addFile" data-key="' + key + '">\n' +
-                '        <img src="' + element.url + '"\n' +
-                '             class="rounded mw-100"\n' +
-                '             style="height: 50px;width: 100%;object-fit: cover;">\n' +
-                '        <p class="text-ellipsis small text-muted mt-1 mb-0" title="' + element.original_name + '">' + element.original_name + '</p>\n' +
-                '    </div>\n' +
-                '</div>';
+            const index = this.page + '-' + key;
+
+            const template = this.element
+                .querySelector('#' + this.data.get('id') + '-media')
+                .content
+                .querySelector('.media-item')
+                .cloneNode(true);
+
+            template.innerHTML = template.innerHTML
+                .replace(/{index}/, index)
+                .replace(/{element.url}/, element.url)
+                .replace(/{element.original_name}/, element.original_name)
+                .replace(/{element.original_name}/, element.original_name);
+
+            $(this.dropname).find(`.media-results`).append(template);
+            this.allMediaList[index] = element;
         });
 
-        $(this.dropname).find(`.media-results`).html(html);
         $(this.dropname).find(`.media-loader`).hide();
         $(this.dropname).find(`.media-results`).show();
     }
@@ -397,7 +424,7 @@ export default class extends ApplicationController {
      */
     addFile(event) {
         const key = event.currentTarget.dataset.key;
-        const file = this.mediaList[key];
+        const file = this.allMediaList[key]
 
         this.addedExistFile(file);