+++ resources/views/fields/upload.blade.php
@@ -1,22 +1,22 @@
 @component($typeForm, get_defined_vars())
     <div
-            data-controller="upload"
-            data-upload-storage="{{$storage ?? 'public'}}"
-            data-upload-name="{{$name}}"
-            data-upload-id="dropzone-{{$id}}"
-            data-upload-data='@json($value)'
-            data-upload-groups="{{$attributes['groups'] ?? ''}}"
-            data-upload-multiple="{{$attributes['multiple']}}"
-            data-upload-parallel-uploads="{{$parallelUploads }}"
-            data-upload-max-file-size="{{$maxFileSize }}"
-            data-upload-max-files="{{$maxFiles}}"
-            data-upload-timeout="{{$timeOut}}"
-            data-upload-accepted-files="{{$acceptedFiles }}"
-            data-upload-resize-quality="{{$resizeQuality }}"
-            data-upload-resize-width="{{$resizeWidth }}"
-            data-upload-is-media-library="{{ $media }}"
-            data-upload-close-on-add="{{ $closeOnAdd }}"
-            data-upload-resize-height="{{$resizeHeight }}"
+        data-controller="upload"
+        data-upload-storage="{{$storage ?? 'public'}}"
+        data-upload-name="{{$name}}"
+        data-upload-id="dropzone-{{$id}}"
+        data-upload-data='@json($value)'
+        data-upload-groups="{{$attributes['groups'] ?? ''}}"
+        data-upload-multiple="{{$attributes['multiple']}}"
+        data-upload-parallel-uploads="{{$parallelUploads }}"
+        data-upload-max-file-size="{{$maxFileSize }}"
+        data-upload-max-files="{{$maxFiles}}"
+        data-upload-timeout="{{$timeOut}}"
+        data-upload-accepted-files="{{$acceptedFiles }}"
+        data-upload-resize-quality="{{$resizeQuality }}"
+        data-upload-resize-width="{{$resizeWidth }}"
+        data-upload-is-media-library="{{ $media }}"
+        data-upload-close-on-add="{{ $closeOnAdd }}"
+        data-upload-resize-height="{{$resizeHeight }}"
     >
         <div id="dropzone-{{$id}}" class="dropzone-wrapper">
             <div class="fallback">
@@ -51,42 +51,43 @@
             <div class="attachment modal fade center-scale" tabindex="-1" role="dialog" aria-hidden="false">
                 <div class="modal-dialog modal-fullscreen-md-down">
                     <div class="modal-content">
-                            <div class="modal-header">
-                                <h4 class="modal-title text-black fw-light">
-                                    {{__('File Information')}}
-                                    <small class="text-muted d-block">{{__('Information to display')}}</small>
-                                </h4>
+                        <div class="modal-header">
+                            <h4 class="modal-title text-black fw-light">
+                                {{__('File Information')}}
+                                <small class="text-muted d-block">{{__('Information to display')}}</small>
+                            </h4>
 
-                                <button type="button" class="btn-close" title="Close" data-bs-dismiss="modal" aria-label="Close">
-                                </button>
+                            <button type="button" class="btn-close" title="Close" data-bs-dismiss="modal"
+                                    aria-label="Close">
+                            </button>
+                        </div>
+                        <div class="modal-body p-4">
+                            <div class="form-group">
+                                <label>{{__('System name')}}</label>
+                                <input type="text" class="form-control" data-target="upload.name" readonly
+                                       maxlength="255">
                             </div>
-                            <div class="modal-body p-4">
-                                <div class="form-group">
-                                    <label>{{__('System name')}}</label>
-                                    <input type="text" class="form-control" data-target="upload.name" readonly
-                                           maxlength="255">
-                                </div>
-                                <div class="form-group">
-                                    <label>{{ __('Display name') }}</label>
-                                    <input type="text" class="form-control" data-target="upload.original"
-                                           maxlength="255" placeholder="{{ __('Display name') }}">
-                                </div>
-                                <div class="form-group">
-                                    <label>{{ __('Alternative text') }}</label>
-                                    <input type="text" class="form-control" data-target="upload.alt"
-                                           maxlength="255" placeholder="{{  __('Alternative text')  }}">
-                                </div>
-                                <div class="form-group">
-                                    <label>{{ __('Description') }}</label>
-                                    <textarea class="form-control no-resize"
-                                              data-target="upload.description"
-                                              placeholder="{{ __('Description') }}"
-                                              maxlength="255"
-                                              rows="3"></textarea>
-                                </div>
+                            <div class="form-group">
+                                <label>{{ __('Display name') }}</label>
+                                <input type="text" class="form-control" data-target="upload.original"
+                                       maxlength="255" placeholder="{{ __('Display name') }}">
+                            </div>
+                            <div class="form-group">
+                                <label>{{ __('Alternative text') }}</label>
+                                <input type="text" class="form-control" data-target="upload.alt"
+                                       maxlength="255" placeholder="{{  __('Alternative text')  }}">
+                            </div>
+                            <div class="form-group">
+                                <label>{{ __('Description') }}</label>
+                                <textarea class="form-control no-resize"
+                                          data-target="upload.description"
+                                          placeholder="{{ __('Description') }}"
+                                          maxlength="255"
+                                          rows="3"></textarea>
+                            </div>
 
 
-                                @if($visibility === 'public')
+                            @if($visibility === 'public')
                                 <div class="form-group">
                                     <a href="#" data-action="click->upload#openLink">
                                         <small>
@@ -96,35 +97,38 @@
                                         </small>
                                     </a>
                                 </div>
-                                @endif
+                            @endif
 
 
-                            </div>
-                            <div class="modal-footer">
-                                <button type="button"
-                                        data-bs-dismiss="modal"
-                                        class="btn btn-link">
+                        </div>
+                        <div class="modal-footer">
+                            <button type="button"
+                                    data-bs-dismiss="modal"
+                                    class="btn btn-link">
                                     <span>
                                         {{__('Close')}}
                                     </span>
-                                </button>
-                                <button type="button" data-action="click->upload#save" class="btn btn-default">
-                                    {{__('Apply')}}
-                                </button>
-                            </div>
+                            </button>
+                            <button type="button" data-action="click->upload#save" class="btn btn-default">
+                                {{__('Apply')}}
+                            </button>
                         </div>
+                    </div>
                 </div>
             </div>
 
-            <div class="media modal fade enter-scale disable-scroll" tabindex="-1" role="dialog" aria-hidden="false">
-                <div class="modal-dialog modal-fullscreen-md-down slide-up">
-                    <div class="modal-content">
+            @if($media)
+                <div class="media modal fade enter-scale disable-scroll" tabindex="-1" role="dialog"
+                     aria-hidden="false">
+                    <div class="modal-dialog modal-dialog-scrollable modal-fullscreen-md-down slide-up">
+                        <div class="modal-content">
                             <div class="modal-header">
                                 <h4 class="modal-title text-black fw-light">
                                     {{__('Media Library')}}
                                     <small class="text-muted d-block">{{__('Previously uploaded files')}}</small>
                                 </h4>
-                                <button type="button" class="btn-close" title="Close" data-bs-dismiss="modal" aria-label="Close">
+                                <button type="button" class="btn-close" title="Close" data-bs-dismiss="modal"
+                                        aria-label="Close">
                                 </button>
                             </div>
                             <div class="modal-body p-4">
@@ -135,24 +139,40 @@
                                             <label>{{__('Search file')}}</label>
                                             <input type="search"
                                                    data-target="upload.search"
-                                                   data-action="keydown->upload#loadMedia"
+                                                   data-action="keydown->upload#resetPage keydown->upload#loadMedia"
                                                    class="form-control"
                                                    placeholder="{{ __('Search...') }}"
                                             >
                                         </div>
 
                                         <div class="media-loader spinner-border" role="status">
-                                            <span class="sr-only">{{ __('Loading...') }}</span>
+                                            <span class="visually-hidden">{{ __('Loading...') }}</span>
                                         </div>
 
-
                                         <div class="row media-results m-0"></div>
+
+                                        <div class="mt-2">
+                                            <button class="btn btn-sm btn-link d-block w-100"
+                                                    data-upload-target="loadmore"
+                                                    data-action="click->upload#loadMore">{{ __('Load more') }}</button>
+                                        </div>
+
                                     </div>
                                 </div>
                             </div>
                         </div>
+                    </div>
                 </div>
-            </div>
+
+                <template id="dropzone-{{$id}}-media">
+                    <div class="col-4 col-sm-3 my-3 position-relative media-item">
+                      <div data-action="click->upload#addFile" data-key="{index}">
+                         <img src="{element.url}" class="rounded mw-100" style="height: 50px;width: 100%;object-fit: cover;">
+                          <p class="text-ellipsis small text-muted mt-1 mb-0" title="{element.original_name}">{element.original_name}</p>
+                        </div>
+                      </div>
+                </template>
+            @endif
 
 
             <template id="dropzone-{{$id}}-remove-button">