+++ resources/views/partials/welcome.blade.php
@@ -1,7 +1,7 @@
 <div class="bg-white rounded-top shadow-sm mb-3">
 
     <div class="row g-0">
-        <div class="col col-lg-7 mt-6 p-4 pe-md-0">
+        <div class="col col-lg-7 mt-6 p-4">
 
             <h2 class="mt-2 text-dark fw-light">
                 Hello, nice to see you!