+++ src/Screen/Repository.php
@@ -18,6 +18,18 @@ class Repository extends \Illuminate\Config\Repository implements Countable
     protected $position = 0;
 
     /**
+     * Create a new configuration repository.
+     *
+     * @param iterable $items
+     *
+     * @return void
+     */
+    public function __construct(iterable $items = [])
+    {
+        $this->items = collect($items)->all();
+    }
+
+    /**
      * @param string     $key
      * @param mixed|null $default
      *