+++ stubs/app/Orchid/Screens/Examples/ExampleScreen.php
@@ -4,7 +4,6 @@ namespace App\Orchid\Screens\Examples;
 
 use App\Orchid\Layouts\Examples\ChartBarExample;
 use App\Orchid\Layouts\Examples\ChartLineExample;
-use App\Orchid\Layouts\Examples\MetricsExample;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Orchid\Screen\Actions\Button;
@@ -28,25 +27,11 @@ class ExampleScreen extends Screen
     erat in luctus.';
 
     /**
-     * Display header name.
-     *
-     * @var string
-     */
-    public $name = 'Example screen';
-
-    /**
-     * Display header description.
-     *
-     * @var string
-     */
-    public $description = 'Sample Screen Components';
-
-    /**
      * Query data.
      *
      * @return array
      */
-    public function query(): array
+    public function query(): iterable
     {
         return [
             'charts'  => [
@@ -80,12 +65,32 @@ class ExampleScreen extends Screen
 
             ],
             'metrics' => [
-                ['keyValue' => number_format(6851, 0), 'keyDiff' => 10.08],
-                ['keyValue' => number_format(24668, 0), 'keyDiff' => -30.76],
-                ['keyValue' => number_format(10000, 0), 'keyDiff' => 0],
-                ['keyValue' => number_format(65661, 2), 'keyDiff' => 3.84],
+                'sales'    => ['value' => number_format(6851), 'diff' => 10.08],
+                'visitors' => ['value' => number_format(24668), 'diff' => -30.76],
+                'orders'   => ['value' => number_format(10000), 'diff' => 0],
+                'total'    => number_format(65661),
             ],
         ];
+    }
+
+    /**
+     * Display header name.
+     *
+     * @return string|null
+     */
+    public function name(): ?string
+    {
+        return 'Example screen';
+    }
+
+    /**
+     * Display header description.
+     *
+     * @return string|null
+     */
+    public function description(): ?string
+    {
+        return 'Sample Screen Components';
     }
 
     /**
@@ -93,7 +98,7 @@ class ExampleScreen extends Screen
      *
      * @return \Orchid\Screen\Action[]
      */
-    public function commandBar(): array
+    public function commandBar(): iterable
     {
         return [
 
@@ -138,10 +143,15 @@ class ExampleScreen extends Screen
      *
      * @return string[]|\Orchid\Screen\Layout[]
      */
-    public function layout(): array
+    public function layout(): iterable
     {
         return [
-            MetricsExample::class,
+            Layout::metrics([
+                'Sales Today'    => 'metrics.sales',
+                'Visitors Today' => 'metrics.visitors',
+                'Pending Orders' => 'metrics.orders',
+                'Total Earnings' => 'metrics.total',
+            ]),
 
             Layout::columns([
                 ChartLineExample::class,